Blink active period box
Note that there needs to be a way to seamlessly kill the animation when the period changes. This can be done with the JQuery .finish() method, but under the current MutationObserver setup I'm afraid the animation will be killed at every update.

The animation also seems to die whenever the clock is manually updated/altered.

POSSIBLE SOLUTION - split the clock and period monitoring into separate MutationObservers?