Force a display screen reload when no. of periods changed
Turns out there was an issue where trying to change
the number of periods didn't update on the display
screen after animations were implemented. jQuery reports
some sort of recursion error, but not what it is. This
should be investigated further.

However, simply forcing a display refresh does the
trick.